core updated

+++ src/UserManager/Rbac/Role/RoleController.php
@@ -236,21 +236,37 @@ class RoleController extends \MagmaCore\Administrator\Controller\AdminController
     protected function unassignPermissionAction(): bool
     {
         /* Get the current role ID cast as an integre */
-        $queriedRoleID = (int)$_GET['role_id'] ?? null;
+        //$queriedRoleID = (int)$_GET['role_id'] ?? null;
+        $queriedRoleID = $this->request->handler()->query->getAlnum('role_id');
+        if ($queriedRoleID === null) {
+            throw new \Exception('Invalid role id pass. This action cannot be completed.');
+        }
         /* Get all the permission assigned to the $queriedRoleID and flatten the array */
         $permissionIDs = $this->flattenArray($this->rolePerm->getRepo()->findBy(['*'], ['role_id' => $queriedRoleID]));
         /* The queried permission to remove from the relationship */
         $queriedPermissionID = $this->thisRouteID();
+
+        /**
+         * permission ID = 117
+         * the $role variable within the html template is returning a different role from the one we deleting permissio from
+         * this is where the issue is coming from. Should pass $this_id to the teemplate as this gets the actual ID of the object
+         * being worked on
+         * 
+         * @todo - look at why their are permission selected in the dropdown list that isn't really selected
+         * @todo users table users registered with deleted_at set to null this needs to be set to 0
+         */
+
         /* Ensure our queried permission is within the list of permissions assigned to the queried role */
         if (in_array($queriedPermissionID, $permissionIDs)) {
+
             foreach ($permissionIDs as $permID) {
                 /* We have an exact match */
                 if ($permID === $queriedPermissionID) {
+                    
                     /* We are adding exact condition to ensure the correct permission is deleted from the queried role */
                     $delete = $this->rolePerm->getRepo()->getEm()->getCrud()->delete(['permission_id' => $queriedPermissionID, 'role_id' => $queriedRoleID]);
                     ($delete === true) ? $this->flashMessage('Permission remove') : $this->flashMessage('Permission failed to unassigned.');
                     return $this->redirect('/admin/role/' . $queriedRoleID . '/assigned');
-
                 }
             }
         }
